+++ pybillomat/invoice_items.py
@@ -115,6 +115,8 @@ class InvoiceItem(Item):
 
         if invoice_item_etree is not None:
             self.load_from_etree(invoice_item_etree)
+        elif id is not None:
+            self.load()
 
 
     @classmethod